Added a message to vpn start script about running as root

+++ scripts/deprecated/vpn_start.py
@@ -1,6 +1,8 @@
 #! /usr/bin/python3
+import os
 import socket
 import subprocess
+import sys
 
 def sh(cmd):
   return subprocess \
@@ -14,6 +16,10 @@ def writeFile(filepath, content):
 
 def getPing(ip):
   return sh('ping -c 1 %s | grep ttl | sed -r \'s/.*time=(\S+).*/\\1/\'' % ip)
+
+if os.geteuid() != 0:
+  print('Must be run as root.')
+  sys.exit()
 
 sh('iptables -F; systemctl stop openvpn@client.service')
 # find the address with the shortest ping:
